+++ src/__generated__/Artwork_artwork.graphql.ts
@@ -1,6 +1,7 @@
 /* tslint:disable */
 
 import { ReaderFragment } from "relay-runtime";
+import { AboutArtist_artwork$ref } from "./AboutArtist_artwork.graphql";
 import { ArtworkActions_artwork$ref } from "./ArtworkActions_artwork.graphql";
 import { ArtworkAvailability_artwork$ref } from "./ArtworkAvailability_artwork.graphql";
 import { ArtworkTombstone_artwork$ref } from "./ArtworkTombstone_artwork.graphql";
@@ -13,7 +14,7 @@ export type Artwork_artwork = {
     readonly images: ReadonlyArray<{
         readonly " $fragmentRefs": ImageCarousel_images$ref;
     } | null> | null;
-    readonly " $fragmentRefs": ArtworkTombstone_artwork$ref & ArtworkActions_artwork$ref & ArtworkAvailability_artwork$ref & SellerInfo_artwork$ref & OtherWorks_artwork$ref;
+    readonly " $fragmentRefs": ArtworkTombstone_artwork$ref & ArtworkActions_artwork$ref & ArtworkAvailability_artwork$ref & SellerInfo_artwork$ref & OtherWorks_artwork$ref & AboutArtist_artwork$ref;
     readonly " $refType": Artwork_artwork$ref;
 };
 
@@ -66,8 +67,13 @@ const node: ReaderFragment = {
       "kind": "FragmentSpread",
       "name": "OtherWorks_artwork",
       "args": null
+    },
+    {
+      "kind": "FragmentSpread",
+      "name": "AboutArtist_artwork",
+      "args": null
     }
   ]
 };
-(node as any).hash = '8fc5f37fde36f7a4c9f075c9e9b3969a';
+(node as any).hash = '230d7d14f29a59369fb609b37c5cc2ad';
 export default node;