+++ src/__generated__/FollowArtistButtonTestsErrorQuery.graphql.ts
@@ -1,0 +1,117 @@
+/* tslint:disable */
+
+import { ConcreteRequest } from "relay-runtime";
+import { FollowArtistButton_artist$ref } from "./FollowArtistButton_artist.graphql";
+export type FollowArtistButtonTestsErrorQueryVariables = {};
+export type FollowArtistButtonTestsErrorQueryResponse = {
+    readonly artist: {
+        readonly " $fragmentRefs": FollowArtistButton_artist$ref;
+    } | null;
+};
+export type FollowArtistButtonTestsErrorQuery = {
+    readonly response: FollowArtistButtonTestsErrorQueryResponse;
+    readonly variables: FollowArtistButtonTestsErrorQueryVariables;
+};
+
+
+
+/*
+query FollowArtistButtonTestsErrorQuery {
+  artist(id: "artistID") {
+    ...FollowArtistButton_artist
+    id
+  }
+}
+
+fragment FollowArtistButton_artist on Artist {
+  gravityID
+  id
+  is_followed
+}
+*/
+
+const node: ConcreteRequest = (function(){
+var v0 = [
+  {
+    "kind": "Literal",
+    "name": "id",
+    "value": "artistID"
+  }
+];
+return {
+  "kind": "Request",
+  "fragment": {
+    "kind": "Fragment",
+    "name": "FollowArtistButtonTestsErrorQuery",
+    "type": "Query",
+    "metadata": null,
+    "argumentDefinitions": [],
+    "selections": [
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "artist",
+        "storageKey": "artist(id:\"artistID\")",
+        "args": (v0/*: any*/),
+        "concreteType": "Artist",
+        "plural": false,
+        "selections": [
+          {
+            "kind": "FragmentSpread",
+            "name": "FollowArtistButton_artist",
+            "args": null
+          }
+        ]
+      }
+    ]
+  },
+  "operation": {
+    "kind": "Operation",
+    "name": "FollowArtistButtonTestsErrorQuery",
+    "argumentDefinitions": [],
+    "selections": [
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "artist",
+        "storageKey": "artist(id:\"artistID\")",
+        "args": (v0/*: any*/),
+        "concreteType": "Artist",
+        "plural": false,
+        "selections": [
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "gravityID",
+            "args": null,
+            "storageKey": null
+          },
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "id",
+            "args": null,
+            "storageKey": null
+          },
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "is_followed",
+            "args": null,
+            "storageKey": null
+          }
+        ]
+      }
+    ]
+  },
+  "params": {
+    "operationKind": "query",
+    "name": "FollowArtistButtonTestsErrorQuery",
+    "id": "87c1d262cef549c6dee1faca31eae886",
+    "text": null,
+    "metadata": {}
+  }
+};
+})();
+(node as any).hash = '97921de2506586348cf2e12ac6f5a01f';
+export default node;